webserver/requirements.txt
setuptools==57.5.0
aniso8601==7.0.0
argcomplete==1.11.1
attrs==19.3.0
bcrypt==3.1.4
boto3==1.17.28
botocore==1.20.28
certifi==2019.11.28
cffi==1.15.1
cfn-flip==1.2.2
chardet==3.0.4
click==7.1.1
cryptography==40.0.2
decorator==4.4.2
docutils==0.15.2
durationpy==0.5
Flask==1.1.2
flask_cors==3.0.10
Flask-Bcrypt==0.7.1
flask-restx==0.5.1
flask-expects-json==1.7.0
Flask-S3==0.3.3
funcy==1.14
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
MarkupSafe==1.1.1
pip-tools==4.5.1
py==1.8.1
pycparser==2.20
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
retry==0.9.2
s3transfer==0.3.3
six==1.14.0
text-unidecode==1.3
toml==0.10.0
tqdm==4.44.1
troposphere==3.0.0
urllib3==1.25.8
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.54.0
zipp==3.1.0
# Flask-OAuthlib==0.9.5
Authlib==0.14.3
Flask-JWT-Extended==4.0.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.6.0
Flask-Login==0.5.0
cachetools==4.2.1
pdfkit==0.6.1
stringcase==1.2.0
Flask-Cors==3.0.10
twilio==7.0.0
SQLAlchemy-serializer==1.4.1
stringcase==1.2.0
xlrd==2.0.1
flask-executor==0.10.0
openpyxl==3.0.9
gunicorn==20.1.0
pymongo==4.1.1
flask-expects-json==1.7.0
ed25519==1.5
PyNaCl==1.5.0
greenlet==2.0.2
eventlet==0.30.2
pydantic==1.10.9
email-validator
pika==1.3.0
Flask-APScheduler==1.12.4
python-logging-loki==0.3.1